backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0

langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.53
langsmith>=0.1.0,<0.2.0
chromadb==0.4.22

PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
Pillow==9.5.0
pytesseract==0.3.10
PyMuPDF==1.24.10
markdown>=3.5,<4
beautifulsoup4>=4.12,<5

httpx==0.25.1
certifi>=2024.0.0
numpy==1.26.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
PyYAML==6.0.1
loguru==0.7.2
tiktoken==0.5.1
tenacity==8.2.3
watchfiles==0.21.0

sentence-transformers==2.5.0
torch==2.1.2
transformers==4.36.2

pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
